+++ src/v3/containers/Portfolio/Portfolio.test.js
@@ -1,10 +1,9 @@
-import { render, fireEvent } from "@testing-library/react";
+import { render, fireEvent, act } from "@testing-library/react";
 import V3Portfolio from "./Portfolio";
 
 jest.mock("../../../components/ChatAgent", () => () => null);
 jest.mock("../../components/HamburgerMenu", () => () => null);
 jest.mock("../../components/NavDots", () => () => null);
-jest.mock("../../components/SwipeHandler", () => () => ({ current: null }));
 
 jest.mock("./theme-provider", () => {
   const React = require("react");
@@ -62,6 +61,63 @@ describe("V3Portfolio wheel navigation", () => {
     Object.defineProperty(scrollable, "scrollTop", { configurable: true, writable: true, value: 700 });
 
     fireEvent.wheel(scrollable, { deltaY: 120 });
+    expect(window.location.hash).toBe("#about");
+  });
+});
+
+function fireSwipeOnElement(element, { startY, endY }) {
+  act(() => {
+    element.dispatchEvent(
+      new TouchEvent("touchstart", {
+        bubbles: true,
+        cancelable: true,
+        touches: [{ clientY: startY }],
+      })
+    );
+  });
+  act(() => {
+    element.dispatchEvent(
+      new TouchEvent("touchend", {
+        bubbles: true,
+        cancelable: true,
+        changedTouches: [{ clientY: endY }],
+      })
+    );
+  });
+}
+
+describe("V3Portfolio touch navigation", () => {
+  beforeEach(() => {
+    window.location.hash = "";
+    jest.useFakeTimers();
+  });
+
+  afterEach(() => {
+    jest.runOnlyPendingTimers();
+    jest.useRealTimers();
+  });
+
+  test("does not navigate section when active scrollable can still scroll", () => {
+    const { getByTestId } = render(<V3Portfolio />);
+    const scrollable = getByTestId("home-scrollable");
+
+    Object.defineProperty(scrollable, "scrollHeight", { configurable: true, value: 1000 });
+    Object.defineProperty(scrollable, "clientHeight", { configurable: true, value: 300 });
+    Object.defineProperty(scrollable, "scrollTop", { configurable: true, writable: true, value: 100 });
+
+    fireSwipeOnElement(scrollable, { startY: 200, endY: 100 });
+    expect(window.location.hash).toBe("");
+  });
+
+  test("navigates to next section when active scrollable is at bottom", () => {
+    const { getByTestId } = render(<V3Portfolio />);
+    const scrollable = getByTestId("home-scrollable");
+
+    Object.defineProperty(scrollable, "scrollHeight", { configurable: true, value: 1000 });
+    Object.defineProperty(scrollable, "clientHeight", { configurable: true, value: 300 });
+    Object.defineProperty(scrollable, "scrollTop", { configurable: true, writable: true, value: 700 });
+
+    fireSwipeOnElement(scrollable, { startY: 200, endY: 100 });
     expect(window.location.hash).toBe("#about");
   });
 });